Add libpeas-based plugins and migrate about: to them
I plan for this to be the future architecture for Sagittarius: a fairly
minimal core with a network of plugins on top.

Note that Builtin is not defined in proto-about.plugin.in because it hid it
from the plugins menu. This isn't great when you're trying to see if the
plugin menu works. Perhaps in future, the Builtin flag could be added back,
but until they are shown unless Hidden is set we can't use them since
peas_gtk_manager_view_set_show_builtin is deprecated.