Only build sources once instead of twice
For the plugins to build, we need to link them against *something* (I
think). To do this, we put them against a shared library that's identical
to the executable. Previously, this was done by compiling everything
*again* 'cause logic, and thus was a little slower. Now, we use
both_libraries which allows for slightly faster compilation.

Regrettably, the minimum Meson version had to be lifted to 0.46.0. I think
this is probably worth it --- the main target I'm aiming for is Ubuntu
20.04 LTS, and since I'm on 19.10 (EOL in a day woot!) with 0.51.2 it's a
safe bet that 20.04 has it. Even Debian stable, as of this commit, has
0.49.2 (and oldstable has backports)!