Add support for client-side certificates
Finally! :tada: This has been in the works for a week or so now, and it
finally seems to be done. I don't like the solution *much*, but it works
and can be refined later.

Specifically, getting rid of GCR should allow a lot of simplification,
and perhaps eventually removal of crypto.c.